// packages/datadog-api-client-v1/models/OrganizationSettings.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AccessRole } from './AccessRole';
import { OrganizationSettingsSaml } from './OrganizationSettingsSaml';
import { OrganizationSettingsSamlAutocreateUsersDomains } from './OrganizationSettingsSamlAutocreateUsersDomains';
import { OrganizationSettingsSamlIdpInitiatedLogin } from './OrganizationSettingsSamlIdpInitiatedLogin';
import { OrganizationSettingsSamlStrictMode } from './OrganizationSettingsSamlStrictMode';
/**
* A JSON array of settings.
*/
export declare class OrganizationSettings {
    /**
    * Whether or not the organization users can share widgets outside of Datadog.
    */
    'privateWidgetShare'?: boolean;
    'saml'?: OrganizationSettingsSaml;
    'samlAutocreateAccessRole'?: AccessRole;
    'samlAutocreateUsersDomains'?: OrganizationSettingsSamlAutocreateUsersDomains;
    /**
    * Whether or not SAML can be enabled for this organization.
    */
    'samlCanBeEnabled'?: boolean;
    /**
    * Identity provider endpoint for SAML authentication.
    */
    'samlIdpEndpoint'?: string;
    'samlIdpInitiatedLogin'?: OrganizationSettingsSamlIdpInitiatedLogin;
    /**
    * Whether or not a SAML identity provider metadata file was provided to the Datadog organization.
    */
    'samlIdpMetadataUploaded'?: boolean;
    /**
    * URL for SAML logging.
    */
    'samlLoginUrl'?: string;
    'samlStrictMode'?: OrganizationSettingsSamlStrictMode;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): OrganizationSettings;
    static serialize(data: OrganizationSettings): {
        [key: string]: any;
    };
    constructor();
}